agents/symptom_extractor.py
#!/usr/bin/env python3
"""
Symptom Extractor Sub-Agent
Parses user text into structured symptoms
"""

import json
from google.adk.agents import Agent


def create_symptom_extractor_agent() -> Agent:
    """
    Create the Symptom Extractor sub-agent.

    Role: Parse user text into structured symptoms.
    Output: JSON with symptoms, model, age, etc.
    """

    symptom_extractor = Agent(
        name="symptom_extractor",
        model="gemini-2.5-flash",
        description="Extracts and structures refrigerator symptoms from user description",
        instruction="""
You are a symptom extraction specialist for Samsung refrigerators.

Your job is to analyze the user's problem description and extract structured information.

## Your Task

Parse the user's message and extract:

1. **Symptoms**: List of specific issues (e.g., "not cooling", "making noise", "leaking water")
2. **Severity**: Critical, High, Medium, Low
3. **Duration**: How long the problem has been occurring
4. **Model Information**: If mentioned (e.g., "Samsung RF28", model number)
5. **Error Codes**: Any error codes displayed
6. **Environmental Factors**: Temperature, location, recent changes

## Output Format

Return ONLY valid JSON (no markdown, no explanations):

```json
{
  "symptoms": [
    {
      "symptom": "not cooling properly",
      "location": "main compartment",
      "severity": "high"
    }
  ],
  "severity_overall": "high",
  "duration": "2 days",
  "model": "Unknown",
  "error_codes": [],
  "environmental_factors": [],
  "structured_query": "refrigerator not cooling in main compartment"
}
```

## Rules

- Be precise and specific
- Use technical terms when appropriate
- Flag safety-critical issues with "critical" severity
- Extract ONLY what the user actually said (don't invent details)
- If information is missing, use "Unknown" or empty arrays

Extract symptoms now.
""",
        tools=[]
    )

    return symptom_extractor


# Export
__all__ = ['create_symptom_extractor_agent']
